+++ packages/leap/lib/leap.dart
@@ -1,5 +1,6 @@
 export 'src/characters/characters.dart';
 export 'src/entities/entities.dart';
+export 'src/ground_tiles/ground_tiles.dart';
 export 'src/input/input.dart';
 export 'src/leap_game.dart';
 export 'src/leap_map.dart';
@@ -10,4 +11,4 @@ export 'src/mixins/mixins.dart';
 export 'src/physical_behaviors/collision/collision.dart';
 export 'src/physical_behaviors/physical_behaviors.dart';
 export 'src/physical_behaviors/velocity/velocity.dart';
-export 'src/tiled_object_factory.dart';
+export 'src/tiled_object_handler.dart';