+++ packages/leap/lib/src/characters/jumper_character.dart
@@ -1,12 +1,10 @@
-import 'package:flame/components.dart';
 import 'package:leap/src/characters/jumper_behavior.dart';
 import 'package:leap/src/entities/entities.dart';
 import 'package:leap/src/leap_game.dart';
 
-class JumperCharacter<TGame extends LeapGame> extends PhysicalEntity<TGame> {
-  JumperCharacter({
-    super.health = 10,
-  }) : super(behaviors: [JumperBehavior()]);
+class JumperCharacter<TGame extends LeapGame> extends Character<TGame> {
+  JumperCharacter({super.removeOnDeath, super.health})
+      : super(behaviors: [JumperBehavior()]);
 
   /// When true the character is facing left, otherwise right.
   bool faceLeft = false;
@@ -31,9 +29,6 @@ class JumperCharacter<TGame extends LeapGame> extends PhysicalEntity<TGame> {
   /// The last ground velocity of the character on the horizontal axis.
   double airXVelocity = 0;
 
-  /// The animation position component of the character.
-  PositionComponent? spriteAnimation;
-
   /// Stop walking.
   void stand() => walking = false;
 
@@ -48,19 +43,12 @@ class JumperCharacter<TGame extends LeapGame> extends PhysicalEntity<TGame> {
   void update(double dt) {
     super.update(dt);
 
-    if (spriteAnimation == null) {
-      return;
-    }
-
-    if (velocity.x < 0) {
-      spriteAnimation!.transform.offset = Vector2(
-        spriteAnimation!.width - width,
-        0,
-      );
-      spriteAnimation!.scale.x = -1;
-    } else if (velocity.x > 0) {
-      spriteAnimation!.transform.offset = Vector2(0, 0);
-      spriteAnimation!.scale.x = 1;
+    if (characterAnimation != null) {
+      if (velocity.x < 0) {
+        characterAnimation!.scale.x = -characterAnimation!.scale.x.abs();
+      } else if (velocity.x > 0) {
+        characterAnimation!.scale.x = characterAnimation!.scale.x.abs();
+      }
     }
   }
 }