+++ packages/leap/lib/src/entities/entities.dart
@@ -1,4 +1,5 @@
-export 'collision_type.dart';
+export 'character.dart';
+export 'character_animation.dart';
 export 'common_tags.dart';
 export 'ladder.dart';
 export 'moving_platform.dart';