+++ packages/leap/lib/src/input/three_button_input.dart
@@ -77,9 +77,6 @@ class ThreeButtonTapInput extends PositionComponent
   TapUpEvent? upEvent;
   TapDownEvent? downEvent;
 
-  @override
-  bool get debugMode => true;
-
   bool get isPressed => downEvent != null && upEvent == null;
 
   bool get isPressedLeft {
@@ -111,9 +108,9 @@ class ThreeButtonTapInput extends PositionComponent
   }
 
   @override
-  Future<void> onLoad() async {
-    size = game.leapMap.size;
-    return super.onLoad();
+  void onGameResize(Vector2 size) {
+    super.onGameResize(size);
+    this.size.setFrom(size);
   }
 
   @override