+++ packages/leap/lib/src/leap_world.dart
@@ -1,3 +1,4 @@
+import 'dart:collection';
 import 'dart:math' as math;
 
 import 'package:flame/components.dart';
@@ -8,7 +9,7 @@ import 'package:leap/leap.dart';
 ///
 /// Any [PhysicalEntity] added anywhere in the [LeapGame] component tree
 /// will automatically be part of the world via [physicals].
-class LeapWorld extends World with HasGameRef<LeapGame> {
+class LeapWorld extends World with HasGameRef<LeapGame>, HasTimeScale {
   LeapWorld({this.tileSize = 16})
       : gravity = tileSize * 32,
         maxVelocity = tileSize * 20;
@@ -28,7 +29,11 @@ class LeapWorld extends World with HasGameRef<LeapGame> {
   void update(double dt) {
     final gAccel = gravity * dt;
     for (final physical in physicals.where(
-      (p) => !p.static && !p.hasStatus<IgnoresGravity>(),
+      (p) =>
+          !p.static &&
+          p.statuses
+              .where((s) => s is IgnoresGravity || s is IgnoredByWorld)
+              .isEmpty,
     )) {
       final y = physical.velocity.y;
       final desiredVelocity = (gAccel * physical.gravityRate) + y;
@@ -37,12 +42,21 @@ class LeapWorld extends World with HasGameRef<LeapGame> {
     super.update(dt);
   }
 
-  /// Returns all the physical entities in the game.
-  Iterable<PhysicalEntity> get physicals {
-    return gameRef
-        .trackedComponents<PhysicalEntity>()
-        .where((p) => !p.isRemoving);
+  final List<PhysicalEntity> _physicals = [];
+  late final _physicalsView = UnmodifiableListView(_physicals);
+
+  /// Called by [PhysicalEntity.onMount]
+  void physicalEntityMounted(PhysicalEntity entity) {
+    _physicals.add(entity);
   }
+
+  /// Called by [PhysicalEntity.onRemove]
+  void physicalEntityRemoved(PhysicalEntity entity) {
+    _physicals.remove(entity);
+  }
+
+  /// Returns all the physical entities in the game.
+  Iterable<PhysicalEntity> get physicals => _physicalsView;
 
   /// Whether or not [other] is outside of the world bounds.
   bool isOutside(PhysicalEntity other) {