+++ packages/leap/lib/src/physical_behaviors/collision/collision_detection_behavior.dart
@@ -1,5 +1,6 @@
 import 'dart:math' as math;
 
+import 'package:flame/components.dart';
 import 'package:leap/leap.dart';
 
 /// Contains all the logic for the collision detection system,
@@ -10,7 +11,10 @@ class CollisionDetectionBehavior extends PhysicalBehavior {
   /// The previous collision information of the entity.
   final CollisionInfo prevCollisionInfo;
 
-  /// Temporal hits list, used to store collision during detection.
+  /// All potential hits for this `update` cycle
+  final List<PhysicalEntity> _potentialHits = [];
+
+  /// Temporary hits list, used to store collision during detection.
   final List<PhysicalEntity> _tmpHits = [];
 
   /// Used to test intersections.
@@ -26,50 +30,42 @@ class CollisionDetectionBehavior extends PhysicalBehavior {
   void update(double dt) {
     super.update(dt);
 
-    if (isRemoving) {
-      return;
-    }
-
-    // NOTE: static entities will never run this behavior, so making entities
-    // static has important for performance
-
     prevCollisionInfo.copyFrom(collisionInfo);
     collisionInfo.reset();
 
-    if (!parent.hasStatus<IgnoresGroundCollisions>()) {
-      groundCollisionDetection(dt);
+    // NOTE: static entities will never run this behavior, so making entities
+    // static is important for performance
+
+    if (isRemoving ||
+        parent.statuses
+            .where((s) => s is IgnoredByWorld || s is IgnoresCollisions)
+            .isNotEmpty) {
+      return;
     }
-    nonMapCollisionDetection(dt);
+
+    _calculatePotentialHits(dt);
+    if (!parent.hasStatus<IgnoresSolidCollisions>()) {
+      _solidCollisionDetection(dt);
+    }
+    if (!parent.hasStatus<IgnoresNonSolidCollisions>()) {
+      _nonSolidCollisionDetection(dt);
+    }
   }
 
-  void nonMapCollisionDetection(double dt) {
-    // This should be a small enough number of  objects that looping over all
-    // of them is efficient (we intentionally don't do this for ground tiles for
-    // that reason)
-
-    // TODO(kurtome): This should probably be changed to track "all collisions"
-    //  instead of just "other collisions".
-    //
-    //  Also, since this happens before VelocityBehavior has updated the
-    //  position (and without the special handling groundCollisionDetection
-    //  does) the collision detection may be one frame later than it should be.
-
-    final nonMapCollidables = world.physicals.where(
-      (other) =>
-          other.collisionType == CollisionType.standard &&
-          // solid collisions are considered during ground detection
-          !parent.isOtherSolid(other),
-    );
-    for (final other in nonMapCollidables) {
-      if (intersects(other)) {
-        collisionInfo.otherCollisions ??= [];
-        collisionInfo.otherCollisions!.add(other);
+  void _nonSolidCollisionDetection(double dt) {
+    // Now that we have up/down/left/right collisions from solid detection phase,
+    // we can re-check which potential hits would still collide even.
+    _proxyHitboxForNonSolidHits();
+    for (final other in _potentialHits) {
+      if (!parent.isOtherSolid(other) && intersectsOther(_hitboxProxy, other)) {
+        collisionInfo.addNonSolidCollision(other);
       }
     }
   }
 
-  /// This handles the tilemap ground tiles collisions
-  void groundCollisionDetection(double dt) {
+  /// This handles the tilemap ground tiles collisions and collisions with any
+  /// other solid entity, as determined by [PhysicalEntity.solidTags]
+  void _solidCollisionDetection(double dt) {
     // The collision detection process work in a few steps:
     //
     // 1. Update x/y position as if there were no collisions (from velocity)
@@ -82,10 +78,10 @@ class CollisionDetectionBehavior extends PhysicalBehavior {
 
     if (velocity.x > 0) {
       // Moving right.
-      _calculateTilemapHits((c) {
+      _calculateSolidHits((c) {
         return c.left <= _hitboxProxy.right &&
             c.right >= _hitboxProxy.right &&
-            !c.tags.contains(CommonTags.platform);
+            c.isSolidFromLeft;
       });
 
       if (_tmpHits.isNotEmpty) {
@@ -94,21 +90,23 @@ class CollisionDetectionBehavior extends PhysicalBehavior {
         if (firstRightHit.isSlopeFromLeft) {
           if (velocity.y >= 0) {
             // Ignore slope underneath while moving upwards.
-            collisionInfo.downCollision = firstRightHit;
+            collisionInfo.addDownCollision(firstRightHit);
           } else {
-            collisionInfo.rightCollision = firstRightHit;
+            collisionInfo.addRightCollision(firstRightHit);
           }
         } else if (firstRightHit.left >= right) {
-          collisionInfo.rightCollision = firstRightHit;
+          _tmpHits
+              .where((c) => c.left == firstRightHit.left)
+              .forEach(collisionInfo.addRightCollision);
         }
       }
     }
     if (velocity.x < 0) {
       // Moving left.
-      _calculateTilemapHits((c) {
+      _calculateSolidHits((c) {
         return c.left <= _hitboxProxy.left &&
             c.right >= _hitboxProxy.left &&
-            !c.tags.contains(CommonTags.platform);
+            c.isSolidFromRight;
       });
 
       if (_tmpHits.isNotEmpty) {
@@ -118,12 +116,14 @@ class CollisionDetectionBehavior extends PhysicalBehavior {
           // Ignore slope underneath while moving upwards, should not collide
           // on left.
           if (velocity.y >= 0) {
-            collisionInfo.downCollision = firstLeftHit;
+            collisionInfo.addDownCollision(firstLeftHit);
           } else {
-            collisionInfo.leftCollision = firstLeftHit;
+            collisionInfo.addLeftCollision(firstLeftHit);
           }
         } else if (firstLeftHit.right <= left) {
-          collisionInfo.leftCollision = firstLeftHit;
+          _tmpHits
+              .where((c) => c.right == firstLeftHit.right)
+              .forEach(collisionInfo.addLeftCollision);
         }
       }
     }
@@ -135,9 +135,14 @@ class CollisionDetectionBehavior extends PhysicalBehavior {
         !collisionInfo.down &&
         !collisionInfo.onSlope) {
       // Moving down.
-      _calculateTilemapHits((c) {
-        return c.bottom >= bottom &&
-            // Bottom edge of this the below top of c.
+      _calculateSolidHits((c) {
+        return c.bottom > bottom &&
+            c.isSolidFromTop &&
+            // For one-way platforms from underneath, make sure this is
+            // currently above it so this doesn't pop up on top of it
+            // when overlapping from the below
+            (c.isSolidFromBottom || c.top >= bottom) &&
+            // Bottom edge of this the below top of c, with current velocity
             c.relativeTop(_hitboxProxy) <= _hitboxProxy.bottom;
       });
 
@@ -150,24 +155,25 @@ class CollisionDetectionBehavior extends PhysicalBehavior {
           }
           return a.relativeTop(_hitboxProxy).compareTo(b.top);
         });
-        final firstBottomHit = _tmpHits.first;
-        collisionInfo.downCollision = firstBottomHit;
+        final firstDown = _tmpHits.first;
+        _tmpHits
+            .where((h) => h.top == firstDown.top)
+            .forEach(collisionInfo.addDownCollision);
       }
     }
 
     if (velocity.y < 0) {
       // Moving up.
-      _calculateTilemapHits((c) {
+      _calculateSolidHits((c) {
         return c.top <= top &&
             // Bottom edge of this the below top of c.
             c.bottom >= _hitboxProxy.top &&
-            !c.tags.contains(CommonTags.platform);
+            c.isSolidFromBottom;
       });
 
       if (_tmpHits.isNotEmpty) {
         _tmpHits.sort((a, b) => a.bottom.compareTo(b.bottom));
-        final firstTopHit = _tmpHits.first;
-        collisionInfo.upCollision = firstTopHit;
+        _tmpHits.forEach(collisionInfo.addUpCollision);
       }
     }
 
@@ -185,9 +191,9 @@ class CollisionDetectionBehavior extends PhysicalBehavior {
         final nextSlope = map.groundTiles[prevDown.gridX + 1]
             [prevDown.gridY + nextSlopeYDelta];
         if (prevDown.right >= left) {
-          collisionInfo.downCollision = prevDown;
+          collisionInfo.addDownCollision(prevDown);
         } else if (nextSlope != null && nextSlope.isSlopeFromRight) {
-          collisionInfo.downCollision = nextSlope;
+          collisionInfo.addDownCollision(nextSlope);
         }
       } else if (velocity.x < 0) {
         // Walking down slope to the left.
@@ -195,12 +201,33 @@ class CollisionDetectionBehavior extends PhysicalBehavior {
         final nextSlope = map.groundTiles[prevDown.gridX - 1]
             [prevDown.gridY + nextSlopeYDelta];
         if (prevDown.left <= right) {
-          collisionInfo.downCollision = prevDown;
+          collisionInfo.addDownCollision(prevDown);
         } else if (nextSlope != null && nextSlope.isSlopeFromLeft) {
-          collisionInfo.downCollision = nextSlope;
+          collisionInfo.addDownCollision(nextSlope);
         }
       }
     }
+  }
+
+  void _proxyHitboxForPotentialHits(double dt) {
+    _hitboxProxy.x = x;
+    _hitboxProxy.y = y;
+    _hitboxProxy.height = height;
+    _hitboxProxy.width = width;
+
+    if (velocity.y > 0) {
+      _hitboxProxy.y = y;
+    } else {
+      _hitboxProxy.y = y + velocity.y * dt;
+    }
+    _hitboxProxy.height = height + velocity.y.abs() * dt;
+
+    if (velocity.x > 0) {
+      _hitboxProxy.x = x;
+    } else {
+      _hitboxProxy.x = x + velocity.x * dt;
+    }
+    _hitboxProxy.width = width + velocity.x.abs() * dt;
   }
 
   void _proxyHitboxForVerticalMovement(double dt) {
@@ -229,8 +256,68 @@ class CollisionDetectionBehavior extends PhysicalBehavior {
     _hitboxProxy.width = width + velocity.x.abs() * dt;
   }
 
-  void _calculateTilemapHits(bool Function(PhysicalEntity) filter) {
+  void _proxyHitboxForNonSolidHits() {
+    _hitboxProxy.x = x;
+    _hitboxProxy.y = y;
+    _hitboxProxy.width = width;
+    _hitboxProxy.height = height;
+
+    // This is intended to run after solid collision detection.
+    // So, up/down/left/right will only be set if this is moving
+    // in that direction and collided with something.
+
+    if (collisionInfo.up) {
+      _hitboxProxy.top = collisionInfo.upCollision!.bottom;
+    }
+    if (collisionInfo.down) {
+      _hitboxProxy.bottom = collisionInfo.downCollision!.top;
+    }
+    if (collisionInfo.left) {
+      _hitboxProxy.left = collisionInfo.leftCollision!.right;
+    }
+    if (collisionInfo.right) {
+      _hitboxProxy.right = collisionInfo.rightCollision!.left;
+    }
+  }
+
+  /// Calculates the solid hits into [_tmpHits] given [filter]
+  void _calculateSolidHits(bool Function(PhysicalEntity) filter) {
     _tmpHits.clear();
+
+    for (final other in _potentialHits) {
+      if (parent.isOtherSolid(other) &&
+          intersectsOther(_hitboxProxy, other) &&
+          filter(other)) {
+        _tmpHits.add(other);
+      }
+    }
+  }
+
+  /// populates [_potentialHits]
+  void _calculatePotentialHits(double dt) {
+    _potentialHits.clear();
+
+    // 1.
+    // Find the world phyiscal entity potential hits.
+
+    _proxyHitboxForPotentialHits(dt);
+    final physicals = world.physicals.where(
+      (p) =>
+          p != parent &&
+          p.statuses
+              .where((s) => s is IgnoredByWorld || s is IgnoredByCollisions)
+              .isEmpty,
+    );
+    for (final other in physicals) {
+      if (!other.isRemoving && intersectsOther(_hitboxProxy, other)) {
+        _potentialHits.add(other);
+      }
+    }
+
+    // 2.
+    // Find the tile map potential hits,
+    // tile map entities are handled specially and not added to the world
+    // because there can be A LOT of tiles.
 
     // Find the edges of the map.
     final maxXTile = map.groundTiles.length - 1;
@@ -245,44 +332,26 @@ class CollisionDetectionBehavior extends PhysicalBehavior {
     for (var j = leftTile; j <= rightTile; j++) {
       for (var i = topTile; i <= bottomTile; i++) {
         final tile = map.groundTiles[j][i];
-        if (tile != null &&
-            intersectsOther(_hitboxProxy, tile) &&
-            tile.collisionType == CollisionType.tilemapGround &&
-            filter(tile)) {
-          _tmpHits.add(tile);
+        if (tile != null) {
+          _potentialHits.add(tile);
         }
-      }
-    }
-
-    // TODO(kurtome): cache this somehow so it's only evaluated once per game
-    //  loop?
-    final nonMapCollidables = world.physicals.where(
-      (p) => p.collisionType == CollisionType.standard,
-    );
-    for (final other in nonMapCollidables) {
-      if (intersectsOther(_hitboxProxy, other) &&
-          parent.isOtherSolid(other) &&
-          filter(other)) {
-        _tmpHits.add(other);
       }
     }
   }
 
   static bool intersectsOther(PhysicalEntity a, PhysicalEntity b) {
+    // https://github.com/kurtome/leap/issues/30
+    // Collision detection math here assumes top left anchor.
+    if (a.anchor != Anchor.topLeft || b.anchor != Anchor.topLeft) {
+      throw AssertionError('Collision detection requires Anchor.topLeft');
+    }
+
     final bHeight = b.bottom - b.relativeTop(a);
     // This works by checking if the distance between the objects is less than
     // their combined width (meaning they must overlap).
     return ((a.centerX - b.centerX).abs() * 2 < (a.width + b.width)) &&
         ((a.centerY - (b.bottom - (bHeight / 2))).abs() * 2 <
             (a.height + bHeight));
-  }
-
-  bool intersects(PhysicalEntity b) {
-    final bHeight = b.bottom - b.relativeTop(parent);
-    // This works by checking if the distance between the objects is less than
-    // their combined width (meaning they must overlap).
-    return ((centerX - b.centerX).abs() * 2 < (width + b.width)) &&
-        ((centerY - (b.bottom - (bHeight / 2))).abs() * 2 < (height + bHeight));
   }
 }
 
@@ -297,4 +366,7 @@ class _HitboxProxyComponent extends PhysicalEntity {
 
   @override
   LeapGame get gameRef => overrideGameRef;
+
+  @override
+  LeapGame get game => overrideGameRef;
 }